Experiment with error handling
In the current example we distinguished two kinds of errors: LogicError
and ChainError, where the latter is basically an invalid transition in
the head state machine and the latter is some unexpected error when
sending a tx to the main chain. The former is thus modelled as Either
and the latter as Exception.

The fact that 'Logic.update' is producing an 'ErrorEffect' is a bit
weird still, maybe a 'Maybe' or 'Either' context on that function fits
better.