Use a generic 'Delay' effect instead to post the fan out
This solution is mostly identical to the previous, with the difference
of being able to constrain on the state in which we post the fan out. We
are not sure whether we would need the constraint of being in
'ClosedState' when posting the fan out (as the closing party), but we
will keep it in for illustrating the benefit of this approach.