Use [Party] instead of Set Party
Especially in 'InitTx' and 'HeadParameters' the order of parties is
important for selecting the leader of a snapshot. Also, the order should
be unique (same for all network participants) and the 'Ord' instance
used by the Haskell 'Set' may not be honoring that, so we rather make it
explicitly a list.