Limit number of parties to commit to 3
Still collectComTx does not validate. Also, we are generating a list
of parties which can contain duplicates and this should not be the
case. It's not a problem given how the validators are written and it
might not be in the long run, but this means there might be some
discrepancy between on and off-chain information.