Pass cardano keys to ensure well-formedness of Init tx
Head script now verifies the number of Initial or Commit inputs
matches the number of parties, so we need to ensure the InitTx
correctly generates the right number of Initial outputs.

This commit also removes an 'error' in the  `observeCommitTx` in favor
of returning Nothing: It seems that a commit tx is structurally close
to an abort tx (it consume an Initial) hence we should simply say that
a transaction consuming more than one Initial is not commit tx and
pass it over to the next observer instead of raising an error.

This points to known shortcomings in our error handling strategy in
the DirectChain module.