Use simpler tx sequence generator
We would like to use the cardano-ledger's generator to test our
internal ledger behaves correctly but this is currently problematic
because it's hard to tweak the generation process to produce only the
kind of txs we want in a Head.

Given we are only using those generators for "smoke" testing the
Ledger's validation function over a sequence of txs, it's probably
fine but we should revisit this in the future.