Replace testnet/vasil-testnet with common instructions in testnets/
The goal of this configuration is to easily access and manage these
public testnet environments. The scripts were migrated to spin up
cardano-node for the right known network and also provides ways to
prepare UTxO to use as fuel + to commit, but also to collapse UTxOs on a
known network back into a single output (was useful in the past after
running some smoke tests).

WIP: Current master requires hydra scripts to be published, which we do
not do yet here. What about the smoke tests in general?