Re-enable forAllSt and observeSomeTx
This should be using the same technique as before. That is, a
HasTransitions type class to enumerate transitions from a starting state
using an existential TransitionFrom type. The ObserveTx type class is to
define what observation code to use for a given transition and the
observeSomeTx does unwrap the existentially quantified
SomeOnChainHeadState and uses both type classes to try out all of these
observe functions.