FIX odd jq installation issues
For one AWS account, we met reproducible jq installation issues:

1. apt-get install jq -y would fail because another process was locking dpkg
2. it would then fail because the package was not found

To solve 1. we add a wait to the dpkg lock so that it does not fail if
some upgrade is running concurrently.

To solve 2. we use apt-get update instead of apt (do not ask me why one is
working and not the other) so that debian repositories are up to date with jq
available for installation.

And while we're at it, we also upgrade the machine before moving forward so that
we have an as up to date as possible system.