Add playground to init investigation.
Added traces to seedFaucet for debugging.
Most of this code will be removed once the issue is fixed.

By increasing the amount of threads to a 100 we observed another
Node.Socket exception.
This means querying the `cardano-node` concurrently may fail for many
reasons and we must handle these exceptions propperly.