Provide seedInput directly from the Wallet
Instead of inferring the seed Input from the UTxO the wallet provides,
using another Wallet function (`getFuelUTxO`), it seems better to
provide the relevant function as part of the Wallet interface.
This also makes it easier to mock when testing, as we can provide an
arbitrary TxIn because it's purpose is just to be used as an
identifier for the Head Id.