Revert "FIX Async issue with tracer"
Unfortunately the Logging change is messing up the logs and that was not captured by the tests.

Not every entry is on a new separate line now. This is because access to stdout is not properly
synchronized between threads and that's what the queue was for.

So we need to bite the bullet and either solve the race condition or use an existing framework (eg. katip)

This reverts commit e7ec259c92718ee10257ef0537aa1d5fa6a59ab0.