Remove useless mutation
The MutateHeadTransition mutation was exercising the case of a rogue
close transaction trying to steal some commits.

The attack could have been the following:
* the head is initialized
* some members commit
* a rogue member creates a close transaction consuming all the commits
  and stealing the corresponding funds.

To explore this scenario, we crafted a close transaction which passes
L2 phase validation and is balanced. So successful in a way.

But we realize that there's no way one could create an open state
(which is needed for a valid close) without first having spent the
commits.

Said differently, the pre-requisite for this mutation was to be in
an impossible state.

Our conclusion is that this mutation does not make sense anymore.
Hence, we remove it.