Enable materialization in our haskell.nix project
This will persist the intermediate results when evaluating the
haskell.nix project in nix.

This means, the nix develop and nix build
derivations are sped up because they do not need to look for the
intermediate derivations.

Also, we expect cache behavior be better as only the end results need to
be cached and fetched between dev machines and CI.

Downside of this is, that any change to the haskell dependencies or
build tools, will need an update of the materialized files. The nix
build will show instructions though.