Documentation can use latest docker image
It's fair to not specify any version of hydra docker image in the
documentation as this will use the latest which is deemed to be
stable.

This will secure the release process and make it smoother as we
won't have to manually update all these files for the next releases.