Improve mutation readability and make explicit its intention.
Chances to have this test produce a verification key equal to the
healthy one is very unlikely.
Anyhow, the changes make the mutation more readable and explicit,
thus the reason for this change.
This mutation also exists in Close so this also brings consistency
to the table between the two.