Migrate to self-hosted runner
We remove the cachix step because:
1. we encounter a bunch of troubles with making it work on a self-hosted runner
2. we don't care about uploading built stuff to cachix from the smoke tests, there are other C.I. steps to do it
3. downloading from cachix works anyway

We remove the github cache because we're going to run the smoke tests several
times on the same self-hosted runner so no need for a cache as everything is
already available locally.

We use a shared absolute path to store the cardano database. We should maybe
move the hydra state to a more local and voltaile directory since we remove
it everytime but that's good enough for now.

We ensure that we do not share the same cardano databases storage between two
concurrent work with concurrency github parameter.