We do not need to wait for a day here
Waiting for a day, at a pace of one block every 20 seconds would
generate 4320 blocks which takes forever to be created and
processes, even in IO sim.

We don't really observe the issue now because we cheat in the
way we generate the block.

This commit helps preparing for a more realistic block generation.