Use exception in Model to signal unexpected SUT behaviour
Exception do not play well with StateT which was an oversight in the
initial design of the `RunModel`. By using `ReaderT` and proper
exception type, we should have better failure reports: io-sim will
catch  those exceptions and therefore dump the logs and execution
trace, whereas now (AFAIK) the error is caught outside of io-sim and
we don't really have much information about the context.