WIP splitting chainState and headState
TODO: update the event queue from the callback without fetching an
      outdated chainState from the headState

PROBLEM: I have no idea how to do this refactoring. We have to forM_:

*  one can observeTx and put events in the queue but will use obsolete chainState from the headState
*  the other one will observeTx and update the local, valid, chainState

Then, next step would be to make the update of queue not dependent on the obsolete
chainState in the headState and only deal with putting stuff in the queue and then,
later on, refactor.

But I don't see any path unless just breaking everything and hack it until it compiles but
I don't want to this. I want to smoothly move the code from where it were to where I want it
to be. Make the change easy and then make the easy change... but how?