Do not go back in time the head in case of rollback.
The reason is after a rollback we will observe a rollforward followed by a tick.
This will contain the next slot on the new adopted forked chain.
This new chain is expected to always be longer than the previous.
Finally remember the new chain has also moved forward in time in terms of slots,
and we expect the new tick to contain a valid next slot number.