Add test checking snapshot contains seen txs
This test is currently failing on purpose. The motivation for this
work comes from the idea of packing multiple transactions in a ReqTx
to guarantee the atomicity of a sequence of txs. Thinking about this
feature, I realised that:

1. To enforce atomicity, we would need the snapshot to contain not
only individual tx but actually the ReqTx, plus the proposed new UTxO
set
2. Our current head logic says nothing about the txs part of a
snapshot beyond that they should be applicable to the current ledger.
3. It's therefore impossible to implement the feature and more
importantly it's super easy for any node to front-run other nodes:
When it's your turn to make a snapshot, just look at the proposed txs
pick and chose the one you wants, and add more applicable txs of your
liking. In the context of a market, or an auction perhaps, this makes
it trivial to cheat as far as I understand it.

Therefore, I think we should augment the protocol to (at least)
require that txs in a snapshot have been previously seen in a ReqTx.