Use a single client per node in benchmarks
A while ago we introduced the `withNewClient` function to submit
transactions in the benchmark because we wanted to increase
concurrency and be able to load the server with more than one
client. We ended up dropping that capability from the benchmarks but
kept having a separate connection between the main body of the
benchmark and the transaction submission process.

This actually introduces a subtle bug that does only surface for
sufficiently large loads. Here is what's happening:
1. The benchmark opens a connection to control the head
2. For each separate list of transaction in the workload, we open a
connection and starts sending txs and reading results
3. Because all Server outputs are broadcast to all clients, the first
connection is also sent the messages but does not read it
4. After a while, which depends on parameters of the machine, the TCP
buffer is full and the TCP protocol triggers a close of the server ->
client part of the connection
5. The client -> server part is still valid so the client can send the
close once the transaction processing ends
6. But as soon as we want to receive something, we start reading stale
messages accumulated in some buffer and then it the closed connection