WIP: Remove tags from http server types
Problems arize when we try to expres the draft request type
in the api.yaml file. How to express a list of tuples?

WIP: Remove tags from http server types

Problems arize when we try to expres the draft request type
in the api.yaml file. How to express a list of tuples?

WIP: continue working on api yaml