Workaround bench timeout
We observe some timeouts on CI when running the benchmarks.

Looking at [this exemple](https://github.com/input-output-hk/hydra/actions/runs/5400717236/jobs/9809622903)
it seems that the bench takes around 600 seconds to finish and
hits the 600 timeout for just a few seconds of writing bench
output to files.

In [this exemple](https://github.com/input-output-hk/hydra/actions/runs/5400717236/jobs/9813272692)
(exact same code as in the previous exemple) we can see the bench
only takes less than a minute.

Not proud because I don't get why we have so much discrepancy but let's push
the timeout to 700 seconds so that we have a more stable CI.