Include invalid txs in the summary
Given the way we generate our datasets we should never observe invalid
txs because 1/ the dataset is a linear sequence of txs consuming the
previous one's output and 2/ we always wait for a tx to be confirmed
before submitting the next one.

However, it seems that under some circumstances, this can happen so we
would like to know when it does. Therefore, we fail the test with a
specific message to give more context about the failure and help
investigate when this happens.