Refactor NewState type to now carry a list of events instead of the head state
This is in order to delay the computation to the next state after persisting the produced events.

Also removed headId from most events types as it is not necessary.
Derive instances for HeadStateEvent.
Added RolledBack HeadStateEvent.

Now the update function produces a list of events for the node to process.