Use helpers to define JSON instances
We always use the 'era' in the types to determine the 'eraProtVerLow'
for decoding. According to the ledger team this is fine.

If we don't have an era, i.e. in types which are only parameterized by
crypto, we use the fixed 'LedgerEra' which we currently target in the
hydra-node.