Drop invalid or non-resolvable txs in MockChain
There is at least the `collectComTx` which is expected to be invalid
when submitted by multiple nodes at the same time (racing to open). By
not failing the model when seeing an unresolvable (txIn already spent)
transaction, we follow the same strategy as employed by the currently
integrated "scriptLedger". That is, just drop these transactions.