Wire heartbeat into reliability
The idea is that the reliability layer works with heartbeats to send
acks which do not change upon a ping, denoting the fact the node
cannot make progress by sending new messages and giving the
opportunity to peers to resend messages not acknowledged because of
transient network failures.

This makes the Reliability highly dependent on Heartbeat which raises
the question of whether or not we would not like to merge the 2.