Stress test peers wait for all messages to be received by both peers
Timeouts are inherently unreliable, esp. given an arbitrary and
unknown list of messages and an arbitrary ordering of actions. Tests
might fail because one of the peers stops before the other and
therefore fails to send Pings which will notify the peer it's missing
message, or fail to take into account the peer's Pings.

This commit replaces complicated timeout logic with a simple STM-based
check that _both_ peers received all the messages.