Make misconfiguration check tests more granular
Checking configuration mismatch raises an error and logs what's wrong
in an end-to-end test is fragile, complicated, and takes a lot of
time. This change replaces a flaky ETE test with several more granular
unit tests that check (and assert) the behaviour of the actual
function invoked.

We should also add a higher-level test, possibly in IOSim that ensures
the startup process of the node actually invokes that check, which
would require moving the some of the logic that starts the node out of
Main file.