Emit DecommitApproved upon seeing the snapshot signed
In case all parties signed the spapsot that contains some
decommit utxo we issue a state change to remove the decommitTx from
existing state and emit DecommitApproved. Don't emit DecommitFinalised
here since we most probably want to do it after the decrementTx is sent
(and observed) on L1.

There is a check to make sure that:
1 Signed spapshot contains some utxo to be decommitted
2 There is existing decommitTx in the state
3 Utxo from the snapshot matches the one present in the state