Fix the observed contesters in Direct chain
Stumbled over this when reviewing another observation function in the
Hydra.Chain.Direct.Tx module.

Obviously the ContestObservation and related types were assuming the
included contesters to be the new list, but the code was looking at the
"old datum".

Did not find or write a test which would catch this (yet).