MockChain only includes valid transactions into blocks
Moved and DRYed usage of scriptLedger and introduced function
'collectTransactions' to emulate the dropping transaction behavior as it
was already intended before (see comment).

As we are checking transactions now on submission and fail there, this
is not really needed, but it is more consistent to include only
transactions in the mock chain "blocks" which were valid at time of
evaluation.