Disable sanchonet integration tests
We currently target and test against cardano-node 8.9.x, while sanchonet
is already requiring 8.10.0-pre which uses different conway-genesis.json
configuration files.

While we have a branch which works against 8.10.0-pre, we want to hold
off on testing and relying on that version until it is supported on all
public networks.