TxTrace: Track decommit UTxO in ModelUTxO
I think I found an easy way to pick something to decommit. I always
generate utxo plus utxo to decommit. When producing the actuall real
snapshot I can just choose to set to mempty one of the fields in the
ModelSnapshot (snapshotUTxO or decommitUTxO) which should give me a
proper subset of something to decommit without any extra changes.

Now is the time to play with constraints in the
arbitraryAction/precondition to make sure we always generate something
sensible.