Use CI devShell (ensure right version of jq) (#1601)
This is fix for a subtle bug brought up in #1338 ; namely, that version
1.6 of `jq` has a bug around integer handling:

```
> nix-shell -p jq -I nixpkgs=https://github.com/NixOS/nixpkgs/archive/976fa3369d722e76f37c77493d99829540d43845.tar.gz
> jq -n '9223372036854775807'
9223372036854776000
```

jq 1.7+ doesn't have this issue. Unfortunately, `ubuntu-latest` ships
with jq 1.6. So we need to use a special devShell.

Happily, [there is this nice GitHub
action](https://github.com/marketplace/actions/nix-develop-action),
which makes it easy!

So, we introduce a devShell with the right version of jq, and run our CI
steps there.