Add onConnectivity to NetworkCallback
Interleaving 'Connectivity' events with 'msg' of a NetworkCallback
'deliver', is fine until we have a component which provides both (e.g.
the 'Etcd') and one a component that requires symmetry between in- and
out-bound messages. For example the 'Authenticate' component cannot be
composed around such a 'Either Connectivity msg' inbound component
because to verify signatures of incoming messages, it would need to
understand that this is an 'Either' and only do it for 'Right msg' ->
this is a leaky abstraction.

This commit intsead does expand 'NetworkCallback' to not only 'deliver',
but also provide an 'onConnectivity' callback. While this makes it not
a (Co-)functor, it is more explicit and allows to thread callbacks easily.