Fix sub-process handling by issuing SIGINT
Turns out that withHydraNode was directly terminating the hydra-node
sub-process with SIGTERM, which is not handled by the Haskell runtime by
default. As we also have no signal handler, graceful shutdown was not
done.

This is now sending SIGINT from the hydra-cluster tests to the
hydra-node, and from the hydra-node to etcd.