add GET /snapshot/latest-seen endpoint (#1886)
<!-- Describe your change here -->

In a situation where the head is not reaching consensus, this endpoint
helps identify the peer not cooperating: missing to sign the current
snapshot in flight (emit AckSn) and preventing its confirmation.

This can happen in multiple situations, such as the one described in
[issue #1773](https://github.com/cardano-scaling/hydra/issues/1773),
often due to a peer going offline or falling out of sync because of
differences in its local ledger state, configuration, or node version.
This leads to conflicting views on UTxO availability, blocking further
snapshot confirmation, and effectively stalling the Hydra head.

---

<!-- Consider each and tick it off one way or the other -->
* [x] CHANGELOG updated or not needed
* [x] Documentation updated or not needed
* [x] Haddocks updated or not needed
* [x] No new TODOs introduced or explained herafter