Move snapshotting of deposits to Tick handler
This should be fairly similar to what we had before, with the difference
snapshots would only be requested after the next Tick (block) seen on
chain.

While this might seem cumbersome now, this is actually what a hydra-node
needs to do and wait even longer before a deposit may be accepted.