Switch to hydra-coding-standards-0.3.0 (#2017)
I think this is a good approach to consistency across project that use
flake-parts:

We introduce a repository `hydra-coding-standards`.

https://github.com/cardano-scaling/hydra-coding-standards

This exposes a flake-parts module that provides an option
`coding.standards.hydra.enable`, that automatically sets all relevant
checks and formatting options. The shell tools are removed, and we rely
on `nix fmt` and the `checks.x86_64-linux.treefmt`.

Also adds `-Werror` checks and `weeder` checks based on the listed
`haskellPackages`.

Also adds `statix` to the checks.

This also means our coding standards has a HEAD and a ChangeLog so the
history of our thinking can be understood through the version history.