Use HeadState to derive headId in Greetings
As we keep a HeadState projection, we can use that directly to gather
the HeadId for the Greetings message. In fact, the documentation was
wrong and suggested that this would be used for the /commit endpoint
while this was not the case and the projection implementation was not
showing the headId in closed states.