Alter how our Blueprint tx works
I realized we could utilize blueprint tx outputs for the 
purpose of depositing in complex scenarios (like reference inputs,
script utxo). This brings in much flexibility and we get to delete 
a lot of code but the user experience will be much better and we reduced 
the amount of bugs we could produce.