Prevent insecure l2 interactions until chain sync (#2290)
<!-- Describe your change here -->

🔑 Motivation

> Closes [#2286](https://github.com/cardano-scaling/hydra/issues/2286)

This PR introduces new synchronization tracking (wall-clock time) and
input-handling logic to improve consistency between the chain component
and node state.

---

🔄 Changes

* `NodeState` has been refactored into a sum type with two variants:
`NodeInSync` and `NodeCatchingUp`.

* A new policy based on the configured `ContestationPeriod` has been
introduced, defining the period of time after which the node is
considered out of sync with the chain. Beyond this period, the node will
refuse to process new transactions or sign snapshots.

* The **head logic** now processes inputs only when the node state's
current slot falls within the safe `ContestationPeriod` window relative
to the latest chain time observed by the chain backend and the current
wall-clock time.

+ **NodeState** transitions between `NodeInSync` and `NodeCatchingUp`
according to the above policy.

  + While in `NodeInSync`, the head logic behaves as usual.

  + While in `NodeCatchingUp`:
      - **Client inputs** are **rejected** if outside the window.
      - This is the starting state.

- **Network inputs** are **re-enqueued** (with TTL) until
synchronization is restored.
⚠️ THIS IS PROBLEMATIC: the network component must be aware of the node
state synced status and accept messages only while `NodeInSync`.

* The **API server** now notifies clients via new server outputs when
the node state falls out of sync with the chain backend and when it
becomes synchronized again.

---

📝 Notes

* We use the wall-clock time and not the latest known tip from the
chain, as it is unreliable while the chain backend is still
synchronizing with the Cardano network.

* A party member cannot go offline for longer than the configured
`ContestationPeriod`; otherwise, it would be at risk, as it would be
unable to contest the head closing. This violates the principle of head
safety.

* We decided to move forward with this _alternative_ solution rather
than the _main proposed one_ because it preserves the node architecture
and allows for fast and easy testing of the protocol’s safety without
requiring major hacks during end-to-end tests. This approach also makes
the protocol safe by design and aware of the passage of time relative to
the latest on-chain observations.
---

<!-- Consider each and tick it off one way or the other -->
* [x] CHANGELOG updated or not needed
* [x] Documentation updated or not needed
* [x] Haddocks updated or not needed
* [x] No new TODOs introduced or explained herafter

---------

Co-authored-by: Sebastian Nagel <sebastian.nagel@ncoding.at>
Co-authored-by: Noon <noon.vandersilk@iohk.io>